General refactor
