+++ src/lib/utils/__tests__/getNewSpot.js
@@ -1,4 +1,4 @@
-import { getNewSpot } from "../getNewSpot";
+import { getNewSpot } from '../spot-utils';
 
 describe("getNewSpot", () => {
   it("should return a correct spot taking into account next spots", () => {
@@ -36,5 +36,18 @@ describe("getNewSpot", () => {
         optimalSpot: availableSpots[0]
       })
     ).toEqual(newSpot);
+  });
+});
+
+describe('getOptimalSpot', () => {
+  it(`
+      when there is no spot, stone must be positioned
+      with left 0 and top to the tallest top spot
+    `, () => {
+    const stone = { width: 70, height: 90 };
+    const availableSpots = [
+      { left: 30, top: 0, right: 100, bottom: null },
+      { left: 0, top: 0, right: 100, bottom: null },
+    ];
   });
 });
