+++ src/lib/utils/__tests__/placeStone.js
@@ -1,4 +1,4 @@
-import { placeStone } from '../placeStone';
+import { placeStone } from '../place-stone';
 
 describe('placeStone', () => {
   it('should place first stone correct', () => {
