+++ src/lib/utils/array-utils.js
@@ -1,4 +1,4 @@
-export function find(func: (item: any) => boolean, list: any[]): any {
+export function find(func, list) {
   for (let i = 0, len = list.length; i < len; i++) {
     const item = list[i];
     const test = func(item);
@@ -9,3 +9,7 @@ export function find(func: (item: any) => boolean, list: any[]): any {
 
   return null;
 }
+
+export const filter = (func, list) => list.filter(func);
+
+export const sort = (func, list) => [ ...list ].sort(func);
