--- src/lib/utils/filter.js
@@ -1,1 +1,0 @@
-export const filter = (func, list) => list.filter(func);
