--- src/lib/utils/sort.js
@@ -1,1 +1,0 @@
-export const sort = (func, list) => [...list].sort(func);
