+++ src/lib/utils/spot-utils.js
@@ -1,3 +1,5 @@
+import { find } from './array-utils';
+
 export const getNewSpot = ({
   availableSpots,
   optimalSpot,
@@ -31,3 +33,14 @@ export const getNewSpot = ({
 
   return newSpot;
 };
+const doesBoxFit = (spot, stone) => {
+  const fitsWidth = spot.right - spot.left >= stone.width;
+  if (!spot.bottom) {
+    return fitsWidth;
+  }
+
+  const fitsHeight = spot.bottom - spot.top >= stone.height;
+  return fitsWidth && fitsHeight;
+};
+export const getOptimalSpot = ({ availableSpots, stone }) =>
+  find(spot => doesBoxFit(spot, stone), availableSpots);
